requirements.txt
requests
beautifulsoup4
biopython
